Fix error when loading in Wayland sessions
Due to changes in GNOME mutter when using Xwayland, GTK may not be
available for extensions at loading time. This results in
Gtk.IconTheme.get_default() returning null and prevents the extension
from loading properly.
This issue is also documented here:
https://gitlab.gnome.org/GNOME/mutter/-/issues/960

This commit creates a custom theme with the gtk_icon_theme
settings value if the get_default() method returns null, as
suggested in the issue linked above.